requirements.txt
streamlit
streamlit-drawable-canvas
pillow
pandas
numpy